pyproject.toml
[project]
name = "asciime"
version = "0.1.0"
description = "Image to ASCII converter (https://alexharri.com/blog/ascii-rendering)"
# readme = "README.md"
requires-python = ">=3.10"
license = "GPL-3.0"
dependencies = [
    "Pillow",
]

authors = [
    { name = "k4lizen", email = "k4lizen@proton.me" }
]

[project.scripts]
asciime = "asciime.main:main"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D101", "D102", "D107", "D103", "D203", "D212", "D203", "D212", "D100", "D104", "EM101", "TRY003", "EM102", "TD001", "TD002", "TD003", "FIX001", "PLR0913", "ERA001", "T201", "PLR2004", "S101"]

[tool.setuptools]
packages = ["asciime"]

[tool.setuptools.package-data]
syzbot_download = ["py.typed"]
